ai-engineering-nexus/contract-engine/pyproject.toml
[project]
name = "contract-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.128.5",
    "fastmcp>=3.0.1",
    "httpx>=0.28.1",
    "jsonschema>=4.26.0",
    "mcp[cli,http]>=1.26.0",
    "openai>=2.17.0",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "redis>=7.1.0",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.40.0",
]
